requirements.txt
libsass
ruff
staticjinja